requirements.txt
pymorphy2==0.8
hpsklearn==0.1.0
pandas==1.1.5
plac==0.9.6
spacy==2.1.9
google==3.0.0
deeppavlov==0.14.0
stanza==1.1.1
liwc==0.5.0
hyperopt==0.2.2
thinc==7.0.8
requests==2.25.1
tabulate==0.8.6
numpy==1.18.0
scipy==1.4.1
wasabi==0.4.0
bert_embedding==1.0.1
xmltodict==0.12.0
textstat==0.6.2
nltk==3.4.5
beautifulsoup4==4.9.3
PyStemmer==2.0.1
scikit_learn==0.24.0
